requirements.txt
-r env/production.txtPackage                 Version  
----------------------- ---------
amqp                    5.0.5    
api                     0.0.7    
asgiref                 3.3.1    
auth                    0.5.3    
autopep8                1.6.0    
billiard                3.6.3.0  
blinker                 1.4      
celery                  5.0.5    
certifi                 2020.12.5
cffi                    1.15.0   
chardet                 4.0.0    
click                   7.1.2    
click-didyoumean        0.0.3    
click-plugins           1.1.1    
click-repl              0.1.6    
coreapi                 2.3.3    
coreschema              0.0.4    
cryptography            36.0.1   
defusedxml              0.7.1    
dj-rest-auth            2.2.2    
Django                  3.1.7    
django-allauth          0.47.0   
django-celery-beat      2.2.0    
django-cors-headers     3.7.0    
django-debug-toolbar    3.2      
django-email-users      1.1.0    
django-extensions       3.1.1    
django-timezone-field   4.1.1    
djangorestframework     3.12.2   
djangorestframework-jwt 1.11.0   
dnspython               2.2.0    
drf-yasg                1.20.0   
eventlet                0.33.0   
falcon                  3.0.1    
greenlet                1.1.2    
gunicorn                20.1.0   
idna                    2.10     
inflection              0.5.1    
itypes                  1.2.0    
Jinja2                  2.11.3   
kombu                   5.0.2    
MarkupSafe              1.1.1    
mongoengine             0.23.1   
nose                    1.3.7    
oauthlib                3.2.0    
packaging               20.9     
pip                     19.2.3   
prompt-toolkit          3.0.16   
psycopg2-binary         2.8.6    
pycodestyle             2.8.0    
pycparser               2.21     
PyJWT                   1.7.1    
pymongo                 3.12.3   
pyparsing               2.4.7    
python-crontab          2.5.1    
python-dateutil         2.8.1    
python-dotenv           0.15.0   
python3-openid          3.2.0    
pytz                    2021.1   
requests                2.25.1   
requests-oauthlib       1.3.1    
ruamel.yaml             0.16.12  
ruamel.yaml.clib        0.2.2    
sentry-sdk              0.20.3   
setuptools              41.2.0   
six                     1.15.0   
social-auth-core        3.4.0    
sqlparse                0.4.1    
toml                    0.10.2   
uritemplate             3.0.1    
urllib3                 1.26.3   
utils                   1.0.1    
vine                    5.0.0    
wcwidth                 0.2.5    
Werkzeug                2.0.2    
Package                  Version  
------------------------ ---------
amqp                     5.0.5    
api                      0.0.7    
asgiref                  3.3.1    
auth                     0.5.3    
autopep8                 1.6.0    
billiard                 3.6.3.0  
blinker                  1.4      
cachetools               5.0.0    
celery                   5.0.5    
certifi                  2020.12.5
cffi                     1.15.0   
chardet                  4.0.0    
click                    7.1.2    
click-didyoumean         0.0.3    
click-plugins            1.1.1    
click-repl               0.1.6    
coreapi                  2.3.3    
coreschema               0.0.4    
cryptography             36.0.1   
defusedxml               0.7.1    
dj-rest-auth             2.2.2    
Django                   3.1.7    
django-allauth           0.47.0   
django-celery-beat       2.2.0    
django-cors-headers      3.7.0    
django-debug-toolbar     3.2      
django-email-users       1.1.0    
django-extensions        3.1.1    
django-storages          1.12.3   
django-timezone-field    4.1.1    
djangorestframework      3.12.2   
djangorestframework-jwt  1.11.0   
dnspython                2.2.0    
drf-yasg                 1.20.0   
eventlet                 0.33.0   
falcon                   3.0.1    
google-api-core          2.5.0    
google-auth              2.6.0    
google-cloud-core        2.2.2    
google-cloud-storage     2.1.0    
google-crc32c            1.3.0    
google-resumable-media   2.1.0    
googleapis-common-protos 1.54.0   
greenlet                 1.1.2    
gunicorn                 20.1.0   
idna                     2.10     
inflection               0.5.1    
itypes                   1.2.0    
Jinja2                   2.11.3   
kombu                    5.0.2    
MarkupSafe               1.1.1    
mongoengine              0.23.1   
nose                     1.3.7    
oauthlib                 3.2.0    
packaging                20.9     
Pillow                   9.0.1    
pip                      19.2.3   
prompt-toolkit           3.0.16   
protobuf                 3.19.4   
psycopg2-binary          2.8.6    
pyasn1                   0.4.8    
pyasn1-modules           0.2.8    
pycodestyle              2.8.0    
pycparser                2.21     
PyJWT                    1.7.1    
pymongo                  3.12.3   
pyparsing                2.4.7    
python-crontab           2.5.1    
python-dateutil          2.8.1    
python-dotenv            0.15.0   
python3-openid           3.2.0    
pytz                     2021.1   
requests                 2.25.1   
requests-oauthlib        1.3.1    
rsa                      4.8      
ruamel.yaml              0.16.12  
ruamel.yaml.clib         0.2.2    
sentry-sdk               0.20.3   
setuptools               41.2.0   
six                      1.15.0   
sqlparse                 0.4.1    
toml                     0.10.2   
uritemplate              3.0.1    
urllib3                  1.26.3   
utils                    1.0.1    
vine                     5.0.0    
wcwidth                  0.2.5    
Werkzeug                 2.0.2    
